Couch view works!
* History API plugged in so that we don't have to reload couch view
lists every time
* Can check 'inclusive' instead of remembering the proper endkey
incantations
