+++ src/components/TypeAhead.svelte
@@ -1,27 +1,14 @@
 <script>
   import { onMount, createEventDispatcher } from "svelte";
   import { state as authState } from "../auth.js";
-  import {
-    databases,
-    documents,
-    design_docs,
-    design_doc_views
-  } from "../couch.js";
+  import { documents } from "../couch.js";
   const dispatch = createEventDispatcher();
 
   let token = $authState.token;
+  let value = "";
 
-  export let value = "";
-
-  export let db = undefined;
-  export let ddoc = undefined;
+  export let id, db;
   export let label = "Please provide a label for this component.";
-  export let mode = db ? "documents" : "databases";
-  let uid = `${mode}.${db ? db : "nil"}.${ddoc ? ddoc : "nil"}`;
-  let datalistId = `data.${uid}`;
-  let inputId = `input.${uid}`;
-
-  console.log(`typeahead initialized with value = ${value}`);
 
   let datalist = [];
 
@@ -32,24 +19,11 @@
   async function lookupIds(event) {
     dispatch("deselected");
     try {
-      if (datalist.length < 1) {
-        if (mode == "databases") {
-          datalist = await databases(token);
-        } else if (mode == "design_doc_views" && db && ddoc) {
-          datalist = Object.keys(await design_doc_views(token, db, ddoc));
-        } else if (mode == "design_docs" && db) {
-          datalist = (await design_docs(token, db)).map(row =>
-            row.id.substring(8)
-          );
-        }
-      }
-      if (mode == "documents") {
-        datalist = (await documents(token, db, {
-          startkey: JSON.stringify(value),
-          endkey: JSON.stringify(`${value}\uFFEF`),
-          limit: 12
-        })).map(row => row.id);
-      }
+      datalist = (await documents(token, db, {
+        startkey: JSON.stringify(value),
+        endkey: JSON.stringify(`${value}\uFFEF`),
+        limit: 12
+      })).map(row => row.id);
     } catch (ignore) {}
   }
 
@@ -59,16 +33,16 @@
 </script>
 
 <div>
-  <label for={inputId}>{label}</label>
+  <label for={`${id}.input`}>{label}</label>
   <input
     type="text"
-    id={inputId}
-    list={datalistId}
+    id={`${id}.input`}
+    list={`${id}.datalist`}
     disabled={datalist.length < 1}
     bind:value
     on:input={lookupIds}
     on:change={selectItem} />
-  <datalist id={datalistId}>
+  <datalist id={`${id}.datalist`}>
     {#each datalist as item}
       <option>{item}</option>
     {/each}
