+++ src/couch.js
@@ -12,29 +12,36 @@ async function _request(token, path, options) {
   return await response.json();
 }
 
-async function databases(token) {
-  return await _request(token, "_all_dbs");
-}
-
 async function documents(token, db, options) {
   let result = await _request(token, [db, "_all_docs"].join("/"), options);
   return result.rows.filter(row => !row.key.startsWith("_design"));
 }
 
-async function design_docs(token, db) {
-  let result = await _request(token, [db, "_all_docs"].join("/"), {
-    startkey: JSON.stringify("_design"),
-    endkey: JSON.stringify("_design\uFFEF")
-  });
-  return result.rows;
-}
-
-async function design_doc_views(token, db, ddoc) {
-  if (ddoc.substring(0, 8) !== "_design") ddoc = `_design/${ddoc}`;
-  return (await _request(token, [db, ddoc].join("/"))).views || {};
+async function design_doc_views(token) {
+  let dbs = await _request(token, "_all_dbs");
+  let views = {};
+  await Promise.all(
+    dbs
+      .filter(db => db[0] !== "_" && db !== "bin")
+      .map(async db => {
+        let ddocs = (await _request(token, [db, "_all_docs"].join("/"), {
+          startkey: JSON.stringify("_design"),
+          endkey: JSON.stringify("_design\uFFEF"),
+          include_docs: true
+        })).rows;
+        ddocs.map(ddoc => {
+          if (ddoc.doc.views) {
+            views[db] = views[db] || {};
+            views[db][ddoc.id.substring(8)] = Object.keys(ddoc.doc.views);
+          }
+        });
+      })
+  );
+  return views;
 }
 
 async function view(token, db, ddoc, view, options) {
+  if (!options.limit) options.limit = 500;
   let result = await _request(
     token,
     [db, "_design", ddoc, "_view", view].join("/"),
@@ -43,4 +50,4 @@ async function view(token, db, ddoc, view, options) {
   return result.rows;
 }
 
-export { databases, documents, design_docs, design_doc_views, view };
+export { documents, design_doc_views, view };
